Implement RAG system with API endpoints for querying and indexing text
