+++ rag/app/models/api.py
@@ -1,0 +1,9 @@
+from pydantic import BaseModel
+from typing import List
+
+class QueryRequest(BaseModel):
+    question: str
+
+class TextIndexRequest(BaseModel):
+    chunks: List[str]
+    source_name: str
